Рефакторинг.
Удалены нереализованные функции, endpoint'ы.
